Telomare

Documentation

class MonoidEndoFolder a where

Methods

monoidFold :: Monoid m => (a -> m) -> a -> m

data IExpr

Constructors

Zero

Pair !IExpr !IExpr

Env

SetEnv !IExpr

Defer !IExpr

Gate !IExpr !IExpr

PLeft !IExpr

PRight !IExpr

Trace

data IExprF r

Constructors

ZeroF

PairF !r !r

EnvF

SetEnvF !r

DeferF !r

GateF !r !r

PLeftF !r

PRightF !r

TraceF

data ExprA a

Constructors

ZeroA a

PairA (ExprA a) (ExprA a) a

EnvA a

SetEnvA (ExprA a) a

DeferA (ExprA a) a

AbortA a

GateA (ExprA a) (ExprA a) a

PLeftA (ExprA a) a

PRightA (ExprA a) a

TraceA a

data ExprT a

Constructors

ZeroT

PairT (ExprT a) (ExprT a)

EnvT

SetEnvT (ExprT a)

DeferT (ExprT a)

AbortT

GateT (ExprT a) (ExprT a)

LeftT (ExprT a)

RightT (ExprT a)

TraceT

TagT (ExprT a) a

getA :: ExprA a -> a

data LamType l

Lambdas can be closed if it's expresion does not depend on any outer binding.

Constructors

Open l

Closed l

data ParserTerm l v

Parser AST

Constructors

TZero

TPair (ParserTerm l v) (ParserTerm l v)

TVar v

TApp (ParserTerm l v) (ParserTerm l v)

TCheck (ParserTerm l v) (ParserTerm l v)

TITE (ParserTerm l v) (ParserTerm l v) (ParserTerm l v)

TLeft (ParserTerm l v)

TRight (ParserTerm l v)

TTrace (ParserTerm l v)

THash (ParserTerm l v)

TChurch Int

TLam (LamType l) (ParserTerm l v)

TLimitedRecursion (ParserTerm l v) (ParserTerm l v) (ParserTerm l v)

data ParserTermF (l :: Type) (v :: Type) r

Constructors

TZeroF

TPairF r r

TVarF v

TAppF r r

TCheckF r r

TITEF r r r

TLeftF r

TRightF r

TTraceF r

THashF r

TChurchF Int

TLamF (LamType l) r

TLimitedRecursionF r r r

indent :: Int -> String -> String

Helper function to indent. Usefull for indented Show instances.

sindent :: String -> State Int String

Indentation with the State Monad.

indentWithOneChild :: String -> State Int String -> State Int String

One child indentation.

indentWithOneChild' :: String -> State Int String -> State Int String

indentWithTwoChildren' :: String -> State Int String -> State Int String -> State Int String

indentWithChildren' :: String -> [State Int String] -> State Int String

indentWithTwoChildren :: String -> State Int String -> State Int String -> State Int String

Two children indentation.

indentWithThreeChildren :: String -> State Int String -> State Int String -> State Int String -> State Int String

dropUntil :: (a -> Bool) -> [a] -> [a]

`dropUntil p xs` drops leading elements until `p $ head xs` is satisfied.

newtype FragIndex

Constructors

FragIndex

Fields

unFragIndex :: Int

data FragExpr a

Constructors

ZeroFrag

PairFrag (FragExpr a) (FragExpr a)

EnvFrag

SetEnvFrag (FragExpr a)

DeferFrag FragIndex

AbortFrag

GateFrag (FragExpr a) (FragExpr a)

LeftFrag (FragExpr a)

RightFrag (FragExpr a)

TraceFrag

AuxFrag a

data FragExprF (a :: Type) r

Constructors

ZeroFragF

PairFragF r r

EnvFragF

SetEnvFragF r

DeferFragF FragIndex

AbortFragF

GateFragF r r

LeftFragF r

RightFragF r

TraceFragF

AuxFragF a

showFragAlg :: Show a => Base (FragExpr a) (State Int String) -> State Int String

newtype EIndex

Constructors

EIndex

Fields

unIndex :: Int

newtype UnsizedRecursionToken

Constructors

UnsizedRecursionToken

Fields

unUnsizedRecursionToken :: Int

data RecursionSimulationPieces a

Constructors

NestedSetEnvs UnsizedRecursionToken

SizingWrapper UnsizedRecursionToken a

data LocTag

Constructors

DummyLoc

Loc Int Int

newtype FragExprUR

Constructors

FragExprUR

Fields

unFragExprUR :: Cofree (FragExprF (RecursionSimulationPieces FragExprUR)) LocTag

type RecursionPieceFrag = RecursionSimulationPieces FragExprUR

type Term1 = Cofree (ParserTermF String String) LocTag

type Term2 = Cofree (ParserTermF () Int) LocTag

newtype Term3

Term3 :: Map FragIndex (FragExpr BreakExtras) -> Term3

Constructors

Term3 (Map FragIndex FragExprUR)

newtype Term4

Constructors

Term4 (Map FragIndex (Cofree (FragExprF Void) LocTag))

type BreakState a b = State (b, FragIndex, Map FragIndex (Cofree (FragExprF a) LocTag))

type BreakState' a b = BreakState a b (Cofree (FragExprF a) LocTag)

type IndExpr = ExprA EIndex

data RunTimeError

Constructors

AbortRunTime IExpr

SetEnvError IExpr

GenericRunTimeError String IExpr

ResultConversionError String

type RunResult = ExceptT RunTimeError IO

class TelomareLike a where

Methods

fromTelomare :: IExpr -> a

toTelomare :: a -> Maybe IExpr

class TelomareLike a => AbstractRunTime a where

Methods

eval :: a -> RunResult a

rootFrag :: Map FragIndex a -> a

zero :: IExpr

pair :: IExpr -> IExpr -> IExpr

var :: IExpr

env :: IExpr

twiddle :: IExpr -> IExpr

app :: IExpr -> IExpr -> IExpr

pleft :: IExpr -> IExpr

pright :: IExpr -> IExpr

setenv :: IExpr -> IExpr

defer :: IExpr -> IExpr

lam :: IExpr -> IExpr

completeLam :: IExpr -> IExpr

ite :: IExpr -> IExpr -> IExpr -> IExpr

varN :: Int -> IExpr

varNF :: Int -> FragExpr a

pattern FirstArg :: IExpr

pattern SecondArg :: IExpr

pattern ThirdArg :: IExpr

pattern FourthArg :: IExpr

pattern Lam :: IExpr -> IExpr

pattern App :: IExpr -> IExpr -> IExpr

pattern TwoArgFun :: IExpr -> IExpr

pattern ITE :: IExpr -> IExpr -> IExpr -> IExpr

pattern EasyTrace :: IExpr -> IExpr

countApps :: Int -> IExpr -> Maybe Int

pattern ChurchNum :: Int -> IExpr

pattern ToChurch :: IExpr

deferF :: Show a => BreakState' a b -> BreakState' a b

pairF :: BreakState' a b -> BreakState' a b -> BreakState' a b

setEnvF :: BreakState' a b -> BreakState' a b

appF :: (Show a, Enum b, Show b) => BreakState' a b -> BreakState' a b -> BreakState' a b

showRunBreakState' :: BreakState' RecursionPieceFrag UnsizedRecursionToken -> String

lamF :: (Show a, Enum b) => BreakState' a b -> BreakState' a b

clamF :: (Show a, Enum b) => BreakState' a b -> BreakState' a b

innerChurchF :: LocTag -> Int -> BreakState' a b

gateF :: BreakState' a b -> BreakState' a b -> BreakState' a b

iteF :: BreakState' a b -> BreakState' a b -> BreakState' a b -> BreakState' a b

repeatFunctionF :: (Show a, Enum b) => LocTag -> FragExpr a -> BreakState' a b

i2cF :: (Show a, Enum b) => LocTag -> Int -> BreakState' a b

unsizedRecursionWrapper :: UnsizedRecursionToken -> BreakState' RecursionPieceFrag UnsizedRecursionToken -> BreakState' RecursionPieceFrag UnsizedRecursionToken -> BreakState' RecursionPieceFrag UnsizedRecursionToken -> BreakState' RecursionPieceFrag UnsizedRecursionToken

nextBreakToken :: (Enum b, Show b) => BreakState a b b

buildFragMap :: BreakState' a b -> Map FragIndex (Cofree (FragExprF a) LocTag)

pattern FirstArgA :: ExprA a

pattern SecondArgA :: ExprA a

pattern ThirdArgA :: ExprA a

pattern FourthArgA :: ExprA a

pattern AppA :: ExprA a -> ExprA a -> ExprA a

pattern LamA :: ExprA a -> ExprA a

pattern TwoArgFunA :: ExprA a -> a -> a -> ExprA a

pattern PlusA :: ExprA a -> ExprA a -> ExprA a

pattern MultA :: ExprA a -> ExprA a -> ExprA a

data DataType

Constructors

ZeroType

ArrType DataType DataType

PairType DataType DataType

data PartialType

Constructors

ZeroTypeP

AnyType

TypeVariable LocTag Int

ArrTypeP PartialType PartialType

PairTypeP PartialType PartialType

mergePairType :: DataType -> DataType

mergePairTypeP :: PartialType -> PartialType

containsFunction :: PartialType -> Bool

cleanType :: PartialType -> Bool

g2i :: IExpr -> Int

i2g :: Int -> IExpr

ints2g :: [Int] -> IExpr

g2Ints :: IExpr -> [Int]

s2g :: String -> IExpr

g2s :: IExpr -> String

toChurch :: Int -> IExpr

i2gF :: LocTag -> Int -> BreakState' a b

ints2gF :: LocTag -> [Int] -> BreakState' a b

s2gF :: LocTag -> String -> BreakState' a b

isNum :: IExpr -> Bool

nextI :: State EIndex EIndex

toIndExpr :: IExpr -> State EIndex IndExpr

toIndExpr' :: IExpr -> IndExpr

telomareToFragmap :: IExpr -> Map FragIndex (Cofree (FragExprF a) LocTag)

fragmapToTelomare :: Map FragIndex (FragExpr a) -> Maybe IExpr

forget :: Corecursive a => Cofree (Base a) anno -> a

tag :: Recursive a => anno -> a -> Cofree (Base a) anno

newtype FragExprURSansAnnotation

Constructors

FragExprURSA

Fields

unFragExprURSA :: FragExpr (RecursionSimulationPieces FragExprURSansAnnotation)

forgetAnnotationFragExprUR :: FragExprUR -> FragExprURSansAnnotation

insertAndGetKey :: (Ord e, Enum e) => a -> State (Map e a) e

pattern AbortRecursion :: IExpr

pattern AbortUser :: IExpr -> IExpr

pattern AbortAny :: IExpr

pattern AbortUnsizeable :: IExpr -> IExpr

data Pattern

AST for patterns in `case` expressions

Constructors

PatternVar String

PatternInt Int

PatternString String

PatternIgnore

PatternPair Pattern Pattern

data PatternF r

Constructors

PatternVarF [Char]

PatternIntF Int

PatternStringF [Char]

PatternIgnoreF

PatternPairF r r

data UnprocessedParsedTerm

Firstly parsed AST sans location annotations

Constructors

VarUP String

ITEUP UnprocessedParsedTerm UnprocessedParsedTerm UnprocessedParsedTerm

LetUP [(String, UnprocessedParsedTerm)] UnprocessedParsedTerm

ListUP [UnprocessedParsedTerm]

IntUP Int

StringUP String

PairUP UnprocessedParsedTerm UnprocessedParsedTerm

AppUP UnprocessedParsedTerm UnprocessedParsedTerm

LamUP String UnprocessedParsedTerm

ChurchUP Int

UnsizedRecursionUP UnprocessedParsedTerm UnprocessedParsedTerm UnprocessedParsedTerm

LeftUP UnprocessedParsedTerm

RightUP UnprocessedParsedTerm

TraceUP UnprocessedParsedTerm

CheckUP UnprocessedParsedTerm UnprocessedParsedTerm

HashUP UnprocessedParsedTerm
On ad hoc user defined types, this term will be substitued to a unique Int.

CaseUP UnprocessedParsedTerm [(Pattern, UnprocessedParsedTerm)]

data UnprocessedParsedTermF r

Constructors

VarUPF [Char]

ITEUPF r r r

LetUPF [([Char], r)] r

ListUPF [r]

IntUPF Int

StringUPF [Char]

PairUPF r r

AppUPF r r

LamUPF [Char] r

ChurchUPF Int

UnsizedRecursionUPF r r r

LeftUPF r

RightUPF r

TraceUPF r

CheckUPF r r

HashUPF r

CaseUPF r [(Pattern, r)]

_CaseUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, [(Pattern, UnprocessedParsedTerm)])

_HashUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_CheckUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_TraceUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_RightUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_LeftUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_UnsizedRecursionUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm, UnprocessedParsedTerm)

_ChurchUP :: Prism' UnprocessedParsedTerm Int

_LamUP :: Prism' UnprocessedParsedTerm (String, UnprocessedParsedTerm)

_AppUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_PairUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_StringUP :: Prism' UnprocessedParsedTerm String

_IntUP :: Prism' UnprocessedParsedTerm Int

_ListUP :: Prism' UnprocessedParsedTerm [UnprocessedParsedTerm]

_LetUP :: Prism' UnprocessedParsedTerm ([(String, UnprocessedParsedTerm)], UnprocessedParsedTerm)

_ITEUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm, UnprocessedParsedTerm)

_VarUP :: Prism' UnprocessedParsedTerm String