Telomare.Parser

Documentation

type AnnotatedUPT = Cofree UnprocessedParsedTermF LocTag

type TelomareParser = Parsec Void String

TelomareParser :: * -> *

parseVariable :: TelomareParser AnnotatedUPT

Parse a variable.

lineComment :: TelomareParser ()

Line comments start with "--".

blockComment :: TelomareParser ()

A block comment starts with "{-" and ends at "-}". Nested block comments are also supported.

sc :: TelomareParser ()

Space Consumer: Whitespace and comment parser that does not consume new-lines.

scn :: TelomareParser ()

Space Consumer: Whitespace and comment parser that does consume new-lines.

lexeme :: TelomareParser a -> TelomareParser a

This is a wrapper for lexemes that picks up all trailing white space using sc

symbol :: String -> TelomareParser String

A parser that matches given text using string internally and then similarly picks up all trailing white space.

reserved :: String -> TelomareParser ()

This is to parse reserved words.

rws :: [String]

List of reserved words

identifier :: TelomareParser String

Variable identifiers can consist of alphanumeric characters, underscore, and must start with an English alphabet letter

parens :: TelomareParser a -> TelomareParser a

Parser for parenthesis.

brackets :: TelomareParser a -> TelomareParser a

Parser for brackets.

curlies :: TelomareParser a -> TelomareParser a

Parser for curly braces

commaSep :: TelomareParser a -> TelomareParser [a]

Comma sepparated TelomareParser that will be useful for lists

integer :: TelomareParser Integer

Integer TelomareParser used by parseNumber and parseChurch

getLineColumn :: ParsecT Void String Identity LocTag

parseString :: TelomareParser AnnotatedUPT

Parse string literal.

parseNumber :: TelomareParser AnnotatedUPT

Parse number (Integer).

parsePair :: TelomareParser AnnotatedUPT

Parse a pair.

parseUnsizedRecursion :: TelomareParser AnnotatedUPT

Parse unsized recursion triple

parseList :: TelomareParser AnnotatedUPT

Parse a list.

parseITE :: TelomareParser AnnotatedUPT

Parse ITE (which stands for "if then else").

parseHash :: TelomareParser AnnotatedUPT

parseCase :: TelomareParser AnnotatedUPT

parseSingleCase :: TelomareParser (Pattern, AnnotatedUPT)

parsePattern :: TelomareParser Pattern

parsePatternPair :: TelomareParser Pattern

parsePatternInt :: TelomareParser Pattern

parsePatternString :: TelomareParser Pattern

parsePatternVar :: TelomareParser Pattern

parsePatternIgnore :: TelomareParser Pattern

parseSingleExpr :: TelomareParser AnnotatedUPT

Parse a single expression.

parseApplied :: TelomareParser AnnotatedUPT

Parse application of functions.

parseLambda :: TelomareParser AnnotatedUPT

Parse lambda expression.

parseSameLvl :: Pos -> TelomareParser a -> TelomareParser a

Parser that fails if indent level is not pos.

parseLet :: TelomareParser AnnotatedUPT

Parse let expression.

parseLongExpr :: TelomareParser AnnotatedUPT

Parse long expression.

parseChurch :: TelomareParser AnnotatedUPT

Parse church numerals (church numerals are a "$" appended to an integer, without any whitespace separation).

parseRefinementCheck :: TelomareParser (AnnotatedUPT -> AnnotatedUPT)

Parse refinement check.

parseAssignment :: TelomareParser (String, AnnotatedUPT)

Parse assignment add adding binding to ParserState.

parseTopLevel :: TelomareParser AnnotatedUPT

Parse top level expressions.

parseTopLevelWithPrelude

Arguments

:: [(String, AnnotatedUPT)]
Prelude

-> TelomareParser AnnotatedUPT

Parse top level expressions.

parseDefinitions :: TelomareParser (AnnotatedUPT -> AnnotatedUPT)

runTelomareParser_ :: Show a => TelomareParser a -> String -> IO ()

Helper function to test parsers without a result.

runTelomareParserWDebug :: Show a => TelomareParser a -> String -> IO ()

Helper function to debug parsers without a result.

runTelomareParser :: Monad m => TelomareParser a -> String -> m a

Helper function to test Telomare parsers with any result.

parseSuccessful :: Monad m => TelomareParser a -> String -> m Bool

Helper function to test if parser was successful.

runParseLongExpr :: String -> Either String UnprocessedParsedTerm

parsePrelude :: String -> Either String [(String, AnnotatedUPT)]

parseOneExprOrTopLevelDefs :: [(String, AnnotatedUPT)] -> TelomareParser AnnotatedUPT

Parse either a single expression or top level definitions defaulting to the main definition. This function is useful and was made for telomare-evaluare

parseWithPrelude

Arguments

:: [(String, AnnotatedUPT)]
Prelude

-> String
Raw string to be parsed

-> Either String AnnotatedUPT
Error on Left

Parse with specified prelude

Orphan instances

Plated UnprocessedParsedTerm

Instance details

Methods

plate :: Traversal' UnprocessedParsedTerm UnprocessedParsedTerm