Telomare.Possible

Documentation

debug :: Bool

debugTrace :: String -> a -> a

anaM' :: forall m t (x :: Type -> Type) a. (Monad m, Corecursive t, x ~ Base t, Traversable x) => (a -> m (Base t a)) -> a -> m t

data PartExprF f

Constructors

ZeroSF

PairSF f f

EnvSF

SetEnvSF f

GateSF f f

LeftSF f

RightSF f

newtype FunctionIndex

Constructors

FunctionIndex

Fields

unFunctionIndex :: Int

data StuckF f

Constructors

DeferSF FunctionIndex f

class BasicBase (g :: Type -> Type) where

Methods

embedB :: PartExprF x -> g x

extractB :: g x -> Maybe (PartExprF x)

class StuckBase (g :: Type -> Type) where

Methods

embedS :: StuckF x -> g x

extractS :: g x -> Maybe (StuckF x)

class SuperBase (g :: Type -> Type) where

Methods

embedP :: SuperPositionF x -> g x

extractP :: g x -> Maybe (SuperPositionF x)

class AbortBase (g :: Type -> Type) where

Methods

embedA :: AbortableF x -> g x

extractA :: g x -> Maybe (AbortableF x)

class UnsizedBase (g :: Type -> Type) where

Methods

embedU :: UnsizedRecursionF x -> g x

extractU :: g x -> Maybe (UnsizedRecursionF x)

pattern BasicFW :: BasicBase g => PartExprF x -> g x

pattern BasicEE :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Recursive g) => PartExprF g -> g

pattern StuckFW :: StuckBase g => StuckF x -> g x

pattern StuckEE :: forall g (f :: Type -> Type). (Base g ~ f, StuckBase f, Recursive g) => StuckF g -> g

pattern SuperFW :: SuperBase g => SuperPositionF x -> g x

pattern SuperEE :: forall g (f :: Type -> Type). (Base g ~ f, SuperBase f, Recursive g) => SuperPositionF g -> g

pattern AbortFW :: AbortBase g => AbortableF x -> g x

pattern AbortEE :: forall g (f :: Type -> Type). (Base g ~ f, AbortBase f, Recursive g) => AbortableF g -> g

pattern UnsizedFW :: UnsizedBase g => UnsizedRecursionF x -> g x

pattern UnsizedEE :: forall g (f :: Type -> Type). (Base g ~ f, UnsizedBase f, Recursive g) => UnsizedRecursionF g -> g

basicEE :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => PartExprF g -> g

stuckEE :: forall g (f :: Type -> Type). (Base g ~ f, StuckBase f, Corecursive g) => StuckF g -> g

superEE :: forall g (f :: Type -> Type). (Base g ~ f, SuperBase f, Corecursive g) => SuperPositionF g -> g

abortEE :: forall g (f :: Type -> Type). (Base g ~ f, AbortBase f, Corecursive g) => AbortableF g -> g

unsizedEE :: forall g (f :: Type -> Type). (Base g ~ f, UnsizedBase f, Corecursive g) => UnsizedRecursionF g -> g

zeroB :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g

pairB :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g -> g -> g

envB :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g

setEnvB :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g -> g

gateB :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g -> g -> g

leftB :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g -> g

rightB :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g -> g

pattern FillFunction :: (Base g ~ f, BasicBase f, Recursive g) => g -> g -> f g

pattern GateSwitch :: (Base g ~ f, BasicBase f, Recursive g) => g -> g -> g -> f g

fillFunction :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g -> g -> g

gateSwitch :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, Corecursive g) => g -> g -> g -> g

basicStep :: (Base g ~ f, BasicBase f, Corecursive g, Recursive g) => (f g -> g) -> f g -> g

basicStepM :: (Base g ~ f, BasicBase f, Traversable f, Corecursive g, Recursive g, PrettyPrintable g, Monad m) => (f (m g) -> m g) -> f (m g) -> m g

transformNoDefer :: (Base g ~ f, StuckBase f, Recursive g) => (f g -> g) -> g -> g

transformNoDeferM :: (Base g ~ f, StuckBase f, Monad m, Recursive g) => (f (m g) -> m g) -> g -> m g

stuckStep :: (Base a ~ f, StuckBase f, BasicBase f, Recursive a, Corecursive a, PrettyPrintable a) => (f a -> a) -> f a -> a

stuckStepM :: (Base a ~ f, Traversable f, StuckBase f, BasicBase f, Recursive a, Corecursive a, PrettyPrintable a, Monad m) => (f (m a) -> m a) -> f (m a) -> m a

evalBottomUp :: forall t (f :: Type -> Type). (Base t ~ f, BasicBase f, StuckBase f, Corecursive t, Recursive t, Recursive t) => (Base t t -> t) -> t -> t

superStep :: (Base a ~ f, BasicBase f, SuperBase f, Recursive a, Corecursive a, PrettyPrintable a) => (a -> a -> a) -> (f a -> a) -> (f a -> a) -> f a -> a

superStepM :: (Base a ~ f, Traversable f, BasicBase f, SuperBase f, Recursive a, Corecursive a, PrettyPrintable a, Monad m) => (a -> a -> a) -> (f (m a) -> m a) -> (f (m a) -> m a) -> f (m a) -> m a

abortStep :: (Base a ~ f, BasicBase f, StuckBase f, AbortBase f, Recursive a, Corecursive a) => (f a -> a) -> f a -> a

abortStepM :: (Base a ~ f, Traversable f, BasicBase f, StuckBase f, AbortBase f, Recursive a, Corecursive a, Monad m) => (f (m a) -> m a) -> f (m a) -> m a

anyFunctionStep :: (Base a ~ f, BasicBase f, SuperBase f, Recursive a, Corecursive a) => (f a -> a) -> f a -> a

anyFunctionStepM :: (Base a ~ f, Traversable f, BasicBase f, SuperBase f, Recursive a, Corecursive a, Monad m) => (f (m a) -> m a) -> f (m a) -> m a

newtype SizedRecursion

Constructors

SizedRecursion

Fields

unSizedRecursion :: Map UnsizedRecursionToken Int

data StrictAccum a x

Constructors

StrictAccum !a x

unsizedStepM :: (Base a ~ f, Traversable f, BasicBase f, StuckBase f, SuperBase f, AbortBase f, UnsizedBase f, Recursive a, Corecursive a, Eq a, PrettyPrintable a) => Int -> (f (StrictAccum SizedRecursion a) -> StrictAccum SizedRecursion a) -> (f (StrictAccum SizedRecursion a) -> StrictAccum SizedRecursion a) -> f (StrictAccum SizedRecursion a) -> StrictAccum SizedRecursion a

data VoidF f

data SuperPositionF f

Constructors

EitherPF !f !f

AnyPF

data AbortableF f

Constructors

AbortF

AbortedF IExpr

mergeBasic :: forall x (f :: Type -> Type). (Base x ~ f, BasicBase f, Eq x, Corecursive x, Recursive x) => (x -> x -> x) -> x -> x -> x

mergeStuck :: forall x (f :: Type -> Type). (Base x ~ f, StuckBase f, Recursive x) => (x -> x -> x) -> x -> x -> x

mergeSuper :: forall x (f :: Type -> Type). (Base x ~ f, SuperBase f, Eq x, Corecursive x, Recursive x) => (x -> x -> x) -> (x -> x -> x) -> x -> x -> x

mergeAbort :: forall x (f :: Type -> Type). (Base x ~ f, AbortBase f, Eq x, Corecursive x, Recursive x) => (x -> x -> x) -> x -> x -> x

mergeUnknown :: forall x (f :: Type -> Type). (Base x ~ f, SuperBase f, Eq x, Corecursive x, Recursive x) => x -> x -> x

data UnsizedRecursionF f

Constructors

RecursionTestF UnsizedRecursionToken f

UnsizedStubF UnsizedRecursionToken f

SizingWrapperF UnsizedRecursionToken f

SizeStageF UnsizedRecursionToken Int f

data StuckExprF f

Constructors

StuckExprB (PartExprF f)

StuckExprS (StuckF f)

type StuckExpr = Fix StuckExprF

data UnsizedExprF f

Constructors

UnsizedExprB (PartExprF f)

UnsizedExprS (StuckF f)

UnsizedExprP (SuperPositionF f)

UnsizedExprA (AbortableF f)

UnsizedExprU (UnsizedRecursionF f)

type UnsizedExpr = Fix UnsizedExprF

data SuperExprF f

Constructors

SuperExprB (PartExprF f)

SuperExprS (StuckF f)

SuperExprA (AbortableF f)

SuperExprP (SuperPositionF f)

type SuperExpr = Fix SuperExprF

data AbortExprF f

Constructors

AbortExprB (PartExprF f)

AbortExprS (StuckF f)

AbortExprA (AbortableF f)

type AbortExpr = Fix AbortExprF

unsized2abortExpr :: UnsizedExpr -> AbortExpr

term3ToUnsizedExpr :: Int -> Term3 -> UnsizedExpr

data SizedResult

Constructors

AbortedSR

UnsizableSR UnsizedRecursionToken

newtype MonoidList a

Constructors

MonoidList

Fields

unMonoidList :: [a]

capMain :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, StuckBase f, SuperBase f, Recursive g, Corecursive g) => g -> g

isClosure :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, StuckBase f, SuperBase f, Recursive g, Corecursive g) => g -> Bool

sizeTerm :: Int -> UnsizedExpr -> Either UnsizedRecursionToken AbortExpr

convertToF :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, StuckBase f, Traversable f, Corecursive g) => IExpr -> g

convertFromF :: forall g (f :: Type -> Type). (Base g ~ f, TelomareLike g, BasicBase f, StuckBase f, Traversable f, Recursive g) => g -> Maybe IExpr

evalBU :: IExpr -> Maybe IExpr

evalBU' :: IExpr -> IO IExpr

term4toAbortExpr :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, StuckBase f, AbortBase f, Corecursive g) => Term4 -> g

abortExprToTerm4 :: forall g (f :: Type -> Type). (Base g ~ f, BasicBase f, StuckBase f, AbortBase f, Foldable f, Recursive g) => g -> Either IExpr Term4

evalA :: (Maybe IExpr -> Maybe IExpr -> Maybe IExpr) -> Maybe IExpr -> Term4 -> Maybe IExpr

Orphan instances

PrettyPrintable Char

Instance details

Methods

showP :: Char -> State Int String