Telomare.Resolver

Documentation

type VarList = [String]

i2t :: Int -> ParserTerm l v

Int to ParserTerm

ints2t :: [Int] -> ParserTerm l v

List of Int's to ParserTerm

s2t :: String -> ParserTerm l v

String to ParserTerm

i2c :: Int -> Term1

Int to Church encoding

findInts :: Pattern -> [UnprocessedParsedTerm -> UnprocessedParsedTerm]

Finds all PatternInt leaves returning "directions" to these leaves through pairs in the form of a combination of RightUP and LeftUP from the root e.g. PatternPair (PatternVar "x") (PatternPair (PatternInt 0) (PatternVar "y")) will return [LeftUP . RightUP]

findStrings :: Pattern -> [UnprocessedParsedTerm -> UnprocessedParsedTerm]

Finds all PatternString leaves returning "directions" to these leaves through pairs in the form of a combination of RightUP and LeftUP from the root e.g. PatternPair (PatternVar "x") (PatternPair (PatternString "Hello, world!") (PatternVar "y")) will return [LeftUP . RightUP]

fitPatternVarsToCasedUPT :: Pattern -> UnprocessedParsedTerm -> UnprocessedParsedTerm

varsUPT :: UnprocessedParsedTerm -> Set String

Collect all free variable names in a UnprocessedParsedTerm expresion

mkLambda4FreeVarUPs :: UnprocessedParsedTerm -> UnprocessedParsedTerm

findPatternVars :: Pattern -> Map String (UnprocessedParsedTerm -> UnprocessedParsedTerm)

pairStructureCheck :: Pattern -> UnprocessedParsedTerm -> UnprocessedParsedTerm

pairRoute2Dirs :: Pattern -> [UnprocessedParsedTerm -> UnprocessedParsedTerm]

pattern2UPT :: Pattern -> UnprocessedParsedTerm

mkCaseAlternative

Arguments

:: UnprocessedParsedTerm
UPT to be cased

-> UnprocessedParsedTerm
Case result to be made lambda and applied

-> Pattern
Pattern

-> UnprocessedParsedTerm
case result as a lambda applied to the appropirate part of the UPT to be cased

case2annidatedIfs

Arguments

:: UnprocessedParsedTerm
Term to be pattern matched

-> [Pattern]
All patterns in a case expression

-> [UnprocessedParsedTerm]
Int leaves as ListUPs on UPT

-> [UnprocessedParsedTerm]
Int leaves as ListUPs on pattern

-> [UnprocessedParsedTerm]
String leaves as ListUPs on UPT

-> [UnprocessedParsedTerm]
String leaves as ListUPs on pattern

-> [UnprocessedParsedTerm]
Case's alternatives

-> UnprocessedParsedTerm

removeCaseUPs :: UnprocessedParsedTerm -> UnprocessedParsedTerm

debruijinize :: MonadFail m => VarList -> Term1 -> m Term2

splitExpr' :: Term2 -> BreakState' RecursionPieceFrag UnsizedRecursionToken

splitExpr :: Term2 -> Term3

makeLambda

Arguments

:: [(String, UnprocessedParsedTerm)]
Bindings

-> String
Variable name

-> Term1
Lambda body

-> Term1

`makeLambda ps vl t1` makes a TLam around t1 with vl as arguments. Automatic recognition of Close or Open type of TLam.

validateVariables

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude

-> UnprocessedParsedTerm

-> Either String Term1

Transformation from UnprocessedParsedTerm to Term1 validating and inlining VarUPs

varsTerm1 :: Term1 -> Set String

Collect all free variable names in a Term1 expresion

optimizeBuiltinFunctions :: UnprocessedParsedTerm -> UnprocessedParsedTerm

generateAllHashes :: Term2 -> Term2

Process an Term2 to have all HashUP replaced by a unique number. The unique number is constructed by doing a SHA1 hash of the Term2 and adding one for all consecutive HashUP's.

addBuiltins :: UnprocessedParsedTerm -> UnprocessedParsedTerm

process

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude

-> UnprocessedParsedTerm

-> Either String Term3

Process an UnprocessedParsedTerm to a Term3 with failing capability.

process2Term2

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude

-> UnprocessedParsedTerm

-> Either String Term2

runTelomareParser2Term2

Arguments

:: TelomareParser UnprocessedParsedTerm
Parser to run

-> [(String, UnprocessedParsedTerm)]
Prelude

-> String
Raw string to be parsed

-> Either String Term2
Error on Left

Helper function to compile to Term2

parseMain

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude: [(VariableName, BindedUPT)]

-> String
Raw string to be parserd.

-> Either String Term3
Error on Left.

Orphan instances

MonadFail (Either String)

Instance details

Methods

fail :: String -> Either String a