Telomare.Parser

Documentation

data UnprocessedParsedTerm

Constructors

VarUP String

ITEUP UnprocessedParsedTerm UnprocessedParsedTerm UnprocessedParsedTerm

LetUP [(String, UnprocessedParsedTerm)] UnprocessedParsedTerm

ListUP [UnprocessedParsedTerm]

IntUP Int

StringUP String

PairUP UnprocessedParsedTerm UnprocessedParsedTerm

AppUP UnprocessedParsedTerm UnprocessedParsedTerm

LamUP String UnprocessedParsedTerm

ChurchUP Int

UnsizedRecursionUP

LeftUP UnprocessedParsedTerm

RightUP UnprocessedParsedTerm

TraceUP UnprocessedParsedTerm

CheckUP UnprocessedParsedTerm UnprocessedParsedTerm

HashUP UnprocessedParsedTerm
On ad hoc user defined types, this term will be substitued to a unique Int. TODO check

data UnprocessedParsedTermF r

Constructors

VarUPF [Char]

ITEUPF r r r

LetUPF [([Char], r)] r

ListUPF [r]

IntUPF Int

StringUPF [Char]

PairUPF r r

AppUPF r r

LamUPF [Char] r

ChurchUPF Int

UnsizedRecursionUPF

LeftUPF r

RightUPF r

TraceUPF r

CheckUPF r r

HashUPF r

_HashUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_CheckUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_TraceUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_RightUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_LeftUP :: Prism' UnprocessedParsedTerm UnprocessedParsedTerm

_UnsizedRecursionUP :: Prism' UnprocessedParsedTerm ()

_ChurchUP :: Prism' UnprocessedParsedTerm Int

_LamUP :: Prism' UnprocessedParsedTerm (String, UnprocessedParsedTerm)

_AppUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_PairUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm)

_StringUP :: Prism' UnprocessedParsedTerm String

_IntUP :: Prism' UnprocessedParsedTerm Int

_ListUP :: Prism' UnprocessedParsedTerm [UnprocessedParsedTerm]

_LetUP :: Prism' UnprocessedParsedTerm ([(String, UnprocessedParsedTerm)], UnprocessedParsedTerm)

_ITEUP :: Prism' UnprocessedParsedTerm (UnprocessedParsedTerm, UnprocessedParsedTerm, UnprocessedParsedTerm)

_VarUP :: Prism' UnprocessedParsedTerm String

type VarList = [String]

type TelomareParser = Parsec Void String

TelomareParser :: * -> * type TelomareParser = State.StateT ParserState (Parsec Void String)

newtype ErrorString

Constructors

MkES

Fields

getErrorString :: String

i2t :: Int -> ParserTerm l v

Int to ParserTerm

ints2t :: [Int] -> ParserTerm l v

List of Int's to ParserTerm

s2t :: String -> ParserTerm l v

String to ParserTerm

i2c :: Int -> Term1

Int to Church encoding

debruijinize :: MonadFail m => VarList -> Term1 -> m Term2

splitExpr' :: Term2 -> BreakState' BreakExtras BreakExtras

splitExpr :: Term2 -> Term3

parseVariable :: TelomareParser UnprocessedParsedTerm

Parse a variable.

lineComment :: TelomareParser ()

Line comments start with "--".

blockComment :: TelomareParser ()

A block comment starts with "{-" and ends at "-}". Nested block comments are also supported.

sc :: TelomareParser ()

Space Consumer: Whitespace and comment parser that does not consume new-lines.

scn :: TelomareParser ()

Space Consumer: Whitespace and comment parser that does consume new-lines.

lexeme :: TelomareParser a -> TelomareParser a

This is a wrapper for lexemes that picks up all trailing white space using sc

symbol :: String -> TelomareParser String

A parser that matches given text using string internally and then similarly picks up all trailing white space.

reserved :: String -> TelomareParser ()

This is to parse reserved words.

rws :: [String]

List of reserved words

identifier :: TelomareParser String

Variable identifiers can consist of alphanumeric characters, underscore, and must start with an English alphabet letter

parens :: TelomareParser a -> TelomareParser a

Parser for parenthesis.

brackets :: TelomareParser a -> TelomareParser a

Parser for brackets.

commaSep :: TelomareParser a -> TelomareParser [a]

Comma sepparated TelomareParser that will be useful for lists

integer :: TelomareParser Integer

Integer TelomareParser used by parseNumber and parseChurch

parseString :: TelomareParser UnprocessedParsedTerm

Parse string literal.

parseNumber :: TelomareParser UnprocessedParsedTerm

Parse number (Integer).

parsePair :: TelomareParser UnprocessedParsedTerm

Parse a pair.

parseList :: TelomareParser UnprocessedParsedTerm

Parse a list.

parseITE :: TelomareParser UnprocessedParsedTerm

Parse ITE (which stands for "if then else").

parseHash :: TelomareParser UnprocessedParsedTerm

parseSingleExpr :: TelomareParser UnprocessedParsedTerm

Parse a single expression.

parseApplied :: TelomareParser UnprocessedParsedTerm

Parse application of functions.

parseLambda :: TelomareParser UnprocessedParsedTerm

Parse lambda expression.

parseSameLvl :: Pos -> TelomareParser a -> TelomareParser a

Parser that fails if indent level is not pos.

parseLet :: TelomareParser UnprocessedParsedTerm

Parse let expression.

parseLongExpr :: TelomareParser UnprocessedParsedTerm

Parse long expression.

parseChurch :: TelomareParser UnprocessedParsedTerm

Parse church numerals (church numerals are a "$" appended to an integer, without any whitespace separation).

parsePartialFix :: TelomareParser UnprocessedParsedTerm

parseRefinementCheck :: TelomareParser (UnprocessedParsedTerm -> UnprocessedParsedTerm)

Parse refinement check.

parseAssignment :: TelomareParser (String, UnprocessedParsedTerm)

Parse assignment add adding binding to ParserState.

parseTopLevel :: TelomareParser UnprocessedParsedTerm

Parse top level expressions.

parseTopLevelWithPrelude

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude

-> TelomareParser UnprocessedParsedTerm

Parse top level expressions.

parseDefinitions :: TelomareParser (UnprocessedParsedTerm -> UnprocessedParsedTerm)

runTelomareParser_ :: Show a => TelomareParser a -> String -> IO ()

Helper function to test parsers without a result.

runTelomareParserWDebug :: Show a => TelomareParser a -> String -> IO ()

Helper function to debug parsers without a result.

runTelomareParser :: Monad m => TelomareParser a -> String -> m a

Helper function to test Telomare parsers with any result.

runTelomareParser2Term2

Arguments

:: TelomareParser UnprocessedParsedTerm
Parser to run

-> [(String, UnprocessedParsedTerm)]
Prelude

-> String
Raw string to be parsed

-> Either String Term2
Error on Left

Helper function to compile to Term2

parseSuccessful :: Monad m => TelomareParser a -> String -> m Bool

Helper function to test if parser was successful.

addBuiltins :: UnprocessedParsedTerm -> UnprocessedParsedTerm

parsePrelude :: String -> Either String [(String, UnprocessedParsedTerm)]

vars :: Term1 -> Set String

Collect all variable names in a Term1 expresion excluding terms binded to lambda args

makeLambda

Arguments

:: [(String, UnprocessedParsedTerm)]
Bindings

-> String
Variable name

-> Term1
Lambda body

-> Term1

`makeLambda ps vl t1` makes a TLam around t1 with vl as arguments. Automatic recognition of Close or Open type of TLam.

validateVariables

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude

-> UnprocessedParsedTerm

-> Either String Term1

Transformation from UnprocessedParsedTerm to Term1 validating and inlining VarUPs

optimizeBuiltinFunctions :: UnprocessedParsedTerm -> UnprocessedParsedTerm

generateAllHashes :: Term2 -> Term2

Process an Term2 to have all HashUP replaced by a unique number. The unique number is constructed by doing a SHA1 hash of the Term2 and adding one for all consecutive HashUP's.

process

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude

-> UnprocessedParsedTerm

-> Either String Term3

Process an UnprocessedParsedTerm to a Term3 with failing capability.

process2Term2

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude

-> UnprocessedParsedTerm

-> Either String Term2

parseWithPrelude

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude

-> String
Raw string to be parsed

-> Either String UnprocessedParsedTerm
Error on Left parseWithPrelude prelude str = bimap errorBundlePretty (LetUP prelude) $ runParser parseTopLevel "" str

Parse with specified prelude

parseMain

Arguments

:: [(String, UnprocessedParsedTerm)]
Prelude: [(VariableName, BindedUPT)]

-> String
Raw string to be parserd.

-> Either String Term3
Error on Left.

Orphan instances

MonadFail (Either String)

Instance details

Methods

fail :: String -> Either String a